+++ CHANGELOG.md
@@ -10,7 +10,11 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 
 ### New
 
+- [#2100](https://gitlab.com/meltano/meltano/-/issues/2100) Let settings be stored directly in `.env` using `meltano config set --store=dotenv`
+
 ### Changes
+
+- [#2100](https://gitlab.com/meltano/meltano/-/issues/2100) Have `meltano ui setup` store secrets in `.env` instead of `ui.cfg`
 
 ### Fixes